 \docType{data}
    \name{best}
    \alias{best}
    \title{The Best Cohort from Maryland}
    \format{a \code{ExpressionSet} instance, 1 row per Affymetrix microarray probe.}
    \source{
      GSE2443
    }
    \description{
      Object imported using GEOquery and curated to remove variables introduced by GEO
    }
    \author{
Mark Dunning
    }
    \usage{varambally}
    \keyword{datasets}